demo projects: support "Primitive Layer"
Since there is no user input and thus no TouchGetMsg(), the general
events are polled during DelayMs(). The delay is not interrupted by the
user, which means if the user initiates quitting the application, it
will not be quit immediately, but after the delay has finished, which
can take up to 5s in this demo.

Antialiased fonts do not work correctly currently, since GetPixel() is
not implemented yet and always returns 0 (black).
There are 1050 invocations of GetPixel() leading to this warning:

WARN: GetPixel() NOT IMPLEMENTED YET