Fix 32-bit integer typedefs in GenericTypeDefs.h of MLA which led to some crashes
They were defined as "signed long" resp. "unsigned long", which was 64 bit on a
64-bit Linux, but 32 bit on a 64-bit Windows 7, regardless whether compiled
with i686-w64-mingw32 or x86_64-w64-mingw32, if native or cross compiled.

This was the reason for the crash in AN1182 of the AppNotes demo, when
switching to Hindi or Thai language.

Another crash occured in the invocation of ExternalMemoryCallback() at the end
of OutCharGetInfoExternal(), where the nCount parameter was huge. (External
memory is not yet supported, but this is an upcoming feature).