SDL: implement CopyBlock() for the renderer
Currently it only works for a single display buffer, srcAddr and dstAddr
are not evaluated.

Only tested with DISP_ORIENTATION=0, not adapted for 90/270 degrees.

SDL_RenderReadPixels() and SDL_RenderDrawPoint() are used for now and
seem to work sufficient regarding performance. Maybe it will later be
converted to SDL_RenderCopy() usage.

SDL_GetWindowPixelFormat() invocations are a real performance killer, happening
e.g. if GetPixel() would have been invoked in the nested for loops
instead of SDL_RenderDrawPoint() directly. A CopyBlock() incovation of a
32x32 px area needed 30-40 ms during these for loops instead of 0-1 ms
if the pixel format has simply been set to RGB888.