SDL: reduce the pixel pitch used in SDL_RenderReadPixels()
It seems to make more sense to set the pitch according to the read
rectangle rather than the whole screen. Then no gaps need to be filled
and if the pixel array wasn't static, some memory could be saved now.

It is also a bit simpler and shorter now and using DISP_HOR_RESOLUTION
as done currently doesn't work correctly with DISP_ORIENTATION set to 90
or 270 anyway.