SDL: simplify scaling handling in CopyBlock()
Do not use a scaled destination rectangle for writing the pixels to the
destination area. Scaling only has to be handled when accessing the
pixels[] array where the read pixel block exists. So the occurrences of
the scaleFactor variable in that area could be reduced from 8 to 3.

The source and destination rectangles now exactly correspond to the
pixels read and written.