SDL: use SDL_RenderCopy() instead of SDL_RenderDrawPoint() in CopyBlock()
Might have better performance compared to the for loop with single pixel
rendering.

A texture has been introduced therefore, only used in CopyBlock().
Currently it is created and destroyed in each invocation, which is
strongly discouraged according to the SDL docs. But at least it only
takes the size of the block to be copied.