SDL: create the texture for CopyBlock() only once during initialization
The size of the texture has to be set to the maximum window size now to
fit all use cases in CopyBlock(). Before, its size was equal to the
block to be copied. In both cases scaling applied.

This change avoids the overhead of creating/destroying the texture
multiple times, but means more RAM usage on the GPU.
But the approach for drawing will be changed anyway soon.