Object Layer demo: disable USE_DOUBLE_BUFFERING
The application is not written for using double buffering, only direct
drawing onto the frame buffer works.

The start animation with the moving Microchip icon and the random fade
effect afterwards did not work anymore. The rest still seems to work
correctly.

"Progress bar" is faster now, when pressing on the ? sign ("How fast can
PIC go?"), but of course it's equally fast as before double buffering
support.